Add alias handling, improve error handling logic

+++ src/modules/events/on_command_error.py
@@ -1,20 +1,48 @@
+import asyncio
 import traceback
+from datetime import datetime
 
-from discord.ext.commands import CheckFailure, CommandError, CommandNotFound, Context, MissingRequiredArgument
+import discord
+from discord.ext.commands import (
+    CheckFailure,
+    CommandError,
+    CommandInvokeError,
+    CommandNotFound,
+    Context,
+    MissingRequiredArgument,
+)
 
+from resources.constants import RED
+from resources.exceptions import HelperError
 from resources.helper_bot import instance as bot
 
 
 @bot.event
 async def on_command_error(ctx: Context, error: CommandError):
-    if (ctx.command is not None and ctx.command.has_error_handler()) or (
-        ctx.cog is not None and ctx.cog.has_error_handler()
-    ):
-        # Ignore commands that have their own error handlers.
+    # For commands with error handlers, we don't handle it here EXCEPT when the error is a HelperError.
+    original_is_custom_error = isinstance(error, CommandInvokeError) and isinstance(
+        error.original, HelperError
+    )
+    valid_error_handler = ctx.command is not None and ctx.command.has_error_handler()
+    valid_cog_handler = ctx.cog is not None and ctx.cog.has_error_handler()
+
+    if (valid_error_handler or valid_cog_handler) and not original_is_custom_error:
         return
+
+    error_embed = discord.Embed(
+        title="<:BloxlinkDead:823633973967716363> Error",
+        description=error,
+        color=RED,
+    )
+    error_embed.set_footer(text="Bloxlink Helper", icon_url=ctx.author.display_avatar)
+    error_embed.timestamp = datetime.now()
+
+    should_delete_original = True
 
     match error:
         case CommandNotFound():
+            # Check if a tag that matches the input can be sent.
+
             name = ctx.invoked_with
             if not name:
                 return
@@ -32,22 +60,13 @@ async def on_command_error(ctx: Context, error: CommandError):
 
                 await bot.db.update_tag(
                     name,
-                    match_command["content"],
                     use_count=match_command["use_count"] + 1,
                 )
 
             return
 
         case CheckFailure():
-            await ctx.reply(
-                content="You do not have permissions to use this command!",
-                mention_author=False,
-                delete_after=5.0,
-                ephemeral=True,
-            )
-
-            if not ctx.interaction:
-                await ctx.message.delete()
+            error_embed.description = "You do not have permissions to use this command!"
 
         case MissingRequiredArgument():
             param = error.param.name
@@ -57,21 +76,36 @@ async def on_command_error(ctx: Context, error: CommandError):
                 param = "the tag content" if param == "tag_content" else "the tag name"
                 message = f"You're missing {param}!"
 
-            await ctx.reply(
-                content=message,
-                mention_author=False,
-                delete_after=5.0,
-                ephemeral=True,
-            )
+            error_embed.description = message
 
-            if not ctx.interaction:
-                await ctx.message.delete()
+        case CommandInvokeError() as err:
+            # Catch HelperError (its wrapped in CommandInvokeError)
+            # Lets us handle the original error if desired.
+            match err.original:
+                case HelperError() as sub_err:
+                    error_embed.description = str(sub_err)
+
+                case _ as sub_err:
+                    error_embed.description = (
+                        f"{sub_err}\n\nAdditional Info:```{traceback.format_exc(chain=True)}```"
+                    )
+                    should_delete_original = False
 
         case _ as err:
-            output = f"{err}\n\nAdditional Info:```{traceback.format_exc(chain=True)}```"
-            await ctx.reply(
-                content=output,
-                mention_author=False,
-                ephemeral=True,
-            )
-            raise err
+            error_embed.description = f"{sub_err}\n\nAdditional Info:```{traceback.format_exc(chain=True)}```"
+            should_delete_original = False
+
+    # Send the error embed
+    await ctx.reply(
+        embed=error_embed,
+        mention_author=False,
+        delete_after=5.0 if should_delete_original else None,
+        ephemeral=True,
+    )
+
+    if not ctx.interaction and should_delete_original:
+        await asyncio.sleep(5)
+        await ctx.message.delete()
+
+    if not should_delete_original:
+        raise err
